+++ index.js
@@ -1,0 +1,4 @@
+"use strict";
+
+module.exports.sql2ast = require('./src/sql2ast.js');
+module.exports.ast2sql = require('./src/ast2sql.js');
